+++ example/src/main.ts
@@ -1,7 +1,8 @@
 import {createI18nInstancePlugin} from "virtual:vue-i18n-types";
 import {createApp} from "vue";
 
-import App from './App.vue'
+import App from '@/App.vue'
+
 
 const app = createApp(App)